# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  resume_dto.CategoryShort:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
  resume_dto.CreateResumeRequest:
    properties:
      adress:
        maxLength: 500
        minLength: 3
        type: string
      category_ids:
        items:
          type: integer
        type: array
      contact:
        maxLength: 255
        minLength: 5
        type: string
      district_id:
        minimum: 1
        type: integer
      experience_year:
        type: integer
      is_active:
        type: boolean
      mahalla_id:
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 2
        type: string
      photo:
        type: string
      price:
        type: integer
      region_id:
        minimum: 1
        type: integer
      skills:
        minLength: 2
        type: string
      text:
        minLength: 10
        type: string
      title:
        maxLength: 500
        minLength: 2
        type: string
    required:
    - adress
    - contact
    - name
    - photo
    - skills
    - text
    - title
    type: object
  resume_dto.ResumeResponse:
    properties:
      adress:
        type: string
      categories:
        items:
          $ref: '#/definitions/resume_dto.CategoryShort'
        type: array
      contact:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      district_id:
        type: integer
      district_name:
        type: string
      experience_year:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      mahalla_id:
        type: integer
      mahalla_name:
        type: string
      name:
        type: string
      photo:
        type: string
      price:
        type: integer
      region_id:
        type: integer
      region_name:
        type: string
      skills:
        type: string
      slug:
        type: string
      text:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      views_count:
        type: integer
    type: object
  resume_dto.UpdateResumeRequest:
    properties:
      adress:
        maxLength: 500
        minLength: 3
        type: string
      category_ids:
        items:
          type: integer
        type: array
      contact:
        maxLength: 255
        minLength: 5
        type: string
      district_id:
        minimum: 1
        type: integer
      experience_year:
        type: integer
      is_active:
        type: boolean
      mahalla_id:
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 2
        type: string
      photo:
        type: string
      price:
        type: integer
      region_id:
        minimum: 1
        type: integer
      skills:
        minLength: 2
        type: string
      text:
        minLength: 10
        type: string
      title:
        maxLength: 500
        minLength: 2
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Admin paneli uchun CRUD endpointlari (cursor pagination, filter, sort)
  title: Usta Backend Admin API
  version: "1.0"
paths:
...
  /resumes:
    get:
      parameters:
      - description: Keyset cursor
        in: query
        name: cursor
        type: string
      - description: Default 10, max 100
        in: query
        name: limit
        type: integer
      - description: id|price|experience_year
        in: query
        name: sort_by
        type: string
      - description: asc|desc
        in: query
        name: sort_order
        type: string
      - description: Name ILIKE
        in: query
        name: name
        type: string
      - description: Title ILIKE
        in: query
        name: title
        type: string
      - description: Name+title+skills+text bo'yicha
        in: query
        name: search
        type: string
      - description: Egasi ID
        in: query
        name: user_id
        type: integer
      - description: Region ID
        in: query
        name: region_id
        type: integer
      - description: District ID
        in: query
        name: district_id
        type: integer
      - description: Mahalla ID
        in: query
        name: mahalla_id
        type: integer
      - description: Holat
        in: query
        name: is_active
        type: boolean
      - description: Minimal narx
        in: query
        name: min_price
        type: integer
      - description: Maksimal narx
        in: query
        name: max_price
        type: integer
      - description: Minimal tajriba (yil)
        in: query
        name: min_experience
        type: integer
      - description: Aniq bitta kategoriya
        in: query
        name: category_id
        type: integer
      - description: Bir nechta kategoriya, vergulli (1,2,3)
        in: query
        name: category_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Resumelar ro'yxati (cursor pagination, public)
      tags:
      - Resumes
    post:
      consumes:
      - application/json
      description: category_ids — bir nechta kategoriya berish mumkin
      parameters:
      - description: Resume
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resume_dto.CreateResumeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resume_dto.ResumeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Resume yaratish (auth user)
      tags:
      - Resumes
  /resumes/{id}:
    delete:
      parameters:
      - description: Resume ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Resumeni o'chirish (egasi yoki admin)
      tags:
      - Resumes
    put:
      consumes:
      - application/json
      description: category_ids berilsa — to'liq almashtiriladi
      parameters:
      - description: Resume ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yangilanadigan maydonlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resume_dto.UpdateResumeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resume_dto.ResumeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Resumeni yangilash (egasi yoki admin)
      tags:
      - Resumes
  /resumes/{slug}:
    get:
      parameters:
      - description: Resume slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resume_dto.ResumeResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resumeni slug bo'yicha olish (public)
      tags:
      - Resumes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"